refactor: mirror user_operator api dir for tests
