+++ bc_obps/registration/tests/endpoints/select_operator/test_request_access.py
@@ -1,0 +1,44 @@
+from model_bakery import baker
+from localflavor.ca.models import CAPostalCodeField
+from registration.models import (
+    User,
+    UserOperator,
+)
+from registration.tests.utils.bakers import (
+    operator_baker,
+)
+from registration.tests.utils.helpers import CommonTestSetup, TestUtils
+from registration.utils import custom_reverse_lazy
+
+
+class TestSelectOperatorRequestAccess(CommonTestSetup):
+    def test_request_subsequent_access_with_valid_payload(self):
+        operator = operator_baker()
+        admin_user = baker.make(User, business_guid=self.user.business_guid)
+        baker.make(
+            UserOperator,
+            user=admin_user,
+            operator=operator,
+            role=UserOperator.Roles.ADMIN,
+            status=UserOperator.Statuses.APPROVED,
+        )
+        response = TestUtils.mock_post_with_auth_role(
+            self,
+            'industry_user',
+            self.content_type,
+            {"operator_id": operator.id},
+            custom_reverse_lazy('request_access'),
+        )
+        response_json = response.json()
+
+        assert response.status_code == 201
+        assert "user_operator_id" in response_json
+
+        user_operator_exists = UserOperator.objects.filter(
+            id=response_json["user_operator_id"],
+            user=self.user,
+            operator=operator,
+            status=UserOperator.Statuses.PENDING,
+        ).exists()
+
+        assert user_operator_exists, "UserOperator object was not created"
