+++ bc_obps/registration/tests/endpoints/user_operator/test_authorization.py
@@ -13,8 +13,6 @@ from registration.tests.utils.bakers import (
 from registration.tests.utils.helpers import CommonTestSetup, TestUtils
 from registration.utils import custom_reverse_lazy
 
-baker.generators.add(CAPostalCodeField, TestUtils.mock_postal_code)
-
 
 class TestUserOperatorEndpointAuthorization(CommonTestSetup):
     def test_user_operator_unauthorized_users_cannot_get(self):
