+++ bc_obps/registration/tests/endpoints/user_operator/test_operator_from_user.py
@@ -1,0 +1,37 @@
+from model_bakery import baker
+from localflavor.ca.models import CAPostalCodeField
+from registration.tests.utils.bakers import (
+    operator_baker,
+)
+from registration.tests.utils.helpers import CommonTestSetup, TestUtils
+from registration.utils import custom_reverse_lazy
+
+
+class TestUserOperatorOperatorFromUserEndpoint(CommonTestSetup):
+    # GET USER OPERATOR OPERATOR ID 200
+    def test_get_user_operator_operator(self):
+        # Act
+        operator = operator_baker()
+        TestUtils.authorize_current_user_as_operator_user(self, operator=operator)
+        response = TestUtils.mock_get_with_auth_role(
+            self, 'industry_user', custom_reverse_lazy('get_user_operator_operator')
+        )
+
+        response_json = response.json()
+
+        # Assert
+        assert response.status_code == 200
+
+        # Additional Assertions
+        assert "operator_id" in response_json
+        assert "status" in response_json
+
+    # GET USER OPERATOR OPERATOR ID 401
+    def test_get_user_operator_operator_with_invalid_user(self):
+        # Act
+        response = TestUtils.mock_get_with_auth_role(
+            self, 'industry_user', custom_reverse_lazy('get_user_operator_operator')
+        )
+
+        # User_operator must be approved to see their operator info
+        assert response.status_code == 401
