+++ bc_obps/registration/tests/endpoints/user_operator/test_user_operator_from_user.py
@@ -1,0 +1,20 @@
+from model_bakery import baker
+from localflavor.ca.models import CAPostalCodeField
+from registration.tests.utils.bakers import (
+    user_operator_baker,
+)
+from registration.tests.utils.helpers import CommonTestSetup, TestUtils
+from registration.utils import custom_reverse_lazy
+
+
+class TestUserOperatorFromUserEndpoint(CommonTestSetup):
+    def test_get_user_operator(self):
+        user_operator = user_operator_baker()
+        user_operator.user_id = self.user.user_guid
+        user_operator.save(update_fields=['user_id'])
+        response = TestUtils.mock_get_with_auth_role(
+            self, 'industry_user', custom_reverse_lazy('get_user_operator_from_user')
+        )
+        assert response.status_code == 200
+        assert response.json()['status'] == user_operator.status
+        assert response.json().get('is_new') is not None
